Add source folders to tailwind.css for scanning
When this layer is extended in a nested manner, all files are not properly scanned to apply tailwind styles. This change forces the lims-layer and lims-app folders to be scanned to resolve this, but creates some coupling between the base layer and layers that extend it (which it shouldn't need to know about). A better solution would keep the layers decoupled, but this is the only thing that has worked so far.